Activity timeline extraction
